refactor: move icy cobble/stone to its own package
