+++ src/main/kotlin/me/grian/griansbetamod/blocks/icystone/IcyCobblestoneBlock.kt
@@ -1,4 +1,4 @@
-package me.grian.griansbetamod.blocks
+package me.grian.griansbetamod.blocks.icystone
 
 import net.minecraft.block.material.Material
 import net.modificationstation.stationapi.api.template.block.TemplateBlock
